+++ grupos/grupo7/001-ejercicios/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["@babel/preset-env"]
+}
