+++ grupos/grupo7/001-ejercicios/.prettierrc
@@ -1,0 +1,3 @@
+{
+  "singleQuote": true
+}
