+++ grupos/grupo7/001-ejercicios/.vscode/settings.json
