+++ grupos/grupo7/001-ejercicios/readme.md
@@ -1,0 +1,31 @@
+# Ejercicios 001
+
+Programar el modulo 'moduloEjercicios.js', de manera de completar las dos funciones que se declaran en ese modulo siguiendo las especificaciones que se encuentran en el comentario de cada funcion.
+
+La idea es utilizar las operaciones de control de flujo que fuimos viendo, como son for, if, y si quieren while o do.
+
+Se pueden crear funciones auxiliares para simplificar o abstraer partes del codigo y mejorar la legibilidad de una funcion, y asi tambien como para reusar el codigo.
+
+## Ejecutando el codigo
+
+Para ejecutar el codigo, hacemos lo siguiente:
+
+- Instalamos las dependencias
+
+```bash
+npm install
+```
+
+- Ejecutamos el punto de entrada de nuestro programa que es index.js
+
+```bash
+npm run start
+```
+
+## Ejercicio
+
+Lo que tenemos que hacer, es implementar las 2 funciones que hay en el archivo moduloEjercicios.js, de manera que utilizen los datos de database.js para producir los valores esperados.
+
+En este caso intencionalemente no hay tests, para que comparemos los resultados que conseguimos dentro del grupo.
+
+No se pueden modificar los datos en basededatos.js
