+++ grupos/grupo7/002-ejercicios/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "001-ejecrcicios",
+  "version": "1.0.0",
+  "description": "",
+  "main": "./src/index.js",
+  "scripts": {
+    "build": "babel ./src/**/*.js -d dist",
+    "start": "npm run build && node dist/index.js",
+    "watch": "nodemon --watch ./src --exec npm run start"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.9.6",
+    "@babel/node": "^7.8.7",
+    "@babel/preset-env": "^7.9.6",
+    "nodemon": "^2.0.4"
+  }
+}
