+++ grupos/grupo7/002-ejercicios/readme.md
@@ -1,0 +1,25 @@
+# Ejercicios 002
+
+Completar las funciones del módulo 'index.js' en el orden en que aparecen.
+
+## Ejecutando el codigo
+
+Para ejecutar el codigo, hacemos lo siguiente:
+
+- Instalamos las dependencias
+
+  ```bash
+  npm install
+  ```
+
+- Para correr el programa
+
+  ```bash
+  npm run start
+  ```
+
+- O con Nodemon
+
+  ```bash
+   npm run watch
+  ```
