+++ grupos/grupo7/002-ejercicios/src/baseDeDatos.js
@@ -2,68 +2,68 @@ export const database = {
   universidades: [
     {
       id: 1,
-      nombre: 'Universidad del Comahue',
+      nombre: "Universidad del Comahue",
       direccion: {
-        calle: 'Av. Siempre viva',
+        calle: "Av. Siempre viva",
         numero: 2043,
-        provincia: 'Neuquen',
+        provincia: "Neuquen",
       },
     },
     {
       id: 2,
-      nombre: 'Universidad de Rio Negro',
-      direccion: { calle: 'Av. Roca', numero: 3023, provincia: 'Rio Negro' },
+      nombre: "Universidad de Rio Negro",
+      direccion: { calle: "Av. Roca", numero: 3023, provincia: "Rio Negro" },
     },
   ],
   profesores: [
-    { id: 1, nombre: 'Jorge Esteban Quito' },
-    { id: 2, nombre: 'Marta Raca' },
-    { id: 3, nombre: 'Silvia Torre Negra' },
-    { id: 4, nombre: 'Ramon Martinez' },
+    { id: 1, nombre: "Jorge Esteban Quito" },
+    { id: 2, nombre: "Marta Raca" },
+    { id: 3, nombre: "Silvia Torre Negra" },
+    { id: 4, nombre: "Ramon Martinez" },
   ],
   materias: [
     {
       id: 1,
-      nombre: 'Análisis matemático',
+      nombre: "Análisis matemático",
       profesores: [1, 2],
       universidad: 1,
     },
     {
       id: 2,
-      nombre: 'Corte y confección de sabanas',
+      nombre: "Corte y confección de sabanas",
       profesores: [3],
       universidad: 2,
     },
     {
       id: 3,
-      nombre: 'Diseño de indumentaria',
+      nombre: "Diseño de indumentaria",
       profesores: [4],
       universidad: 2,
     },
     {
       id: 4,
-      nombre: 'Programación orientada a objetos',
+      nombre: "Programación orientada a objetos",
       profesores: [1, 3],
       universidad: 1,
     },
     {
       id: 5,
-      nombre: 'Ciencias Sociales',
+      nombre: "Ciencias Sociales",
       profesores: [4],
       universidad: 1,
     },
   ],
   provincias: [
-    { id: 1, nombre: 'Neuquen' },
-    { id: 2, nombre: 'Rio Negro' },
+    { id: 1, nombre: "Neuquen" },
+    { id: 2, nombre: "Rio Negro" },
   ],
   alumnos: [
-    { id: 1, nombre: 'Rigoberto Manchu', edad: 22, provincia: 1 },
-    { id: 2, nombre: 'Alina Robles', edad: 21, provincia: 2 },
-    { id: 3, nombre: 'Suzana Mendez', edad: 33, provincia: 2 },
-    { id: 4, nombre: 'Adrian Soto', edad: 26, provincia: 1 },
-    { id: 5, nombre: 'Martin Sarnaga', edad: 23, provincia: 2 },
-    { id: 6, nombre: 'Pablo Tomafi', edad: 30, provincia: 2 },
+    { id: 1, nombre: "Rigoberto Manchu", edad: 22, provincia: 1 },
+    { id: 2, nombre: "Alina Robles", edad: 21, provincia: 2 },
+    { id: 3, nombre: "Suzana Mendez", edad: 33, provincia: 2 },
+    { id: 4, nombre: "Adrian Soto", edad: 26, provincia: 1 },
+    { id: 5, nombre: "Martin Sarnaga", edad: 23, provincia: 2 },
+    { id: 6, nombre: "Pablo Tomafi", edad: 30, provincia: 2 },
   ],
   calificaciones: [
     { alumno: 1, materia: 1, nota: 5 },
@@ -81,5 +81,3 @@ export const database = {
     { alumno: 5, materia: 4, nota: 1 },
   ],
 };
-
-export default database;
